move H160 deserializer to ef_tests/serde_utils.rs
