+++ ef_tests/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod serde_utils;
 pub mod types;
